pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-settlement-dev"
version = "0.9.0"
description = "Dev workspace for a2a-settlement repo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
exchange = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "gunicorn>=21.2",
  "sqlalchemy>=2.0",
  "bcrypt>=4.0",
  "httpx>=0.27",
  "PyNaCl>=1.5.0",
  "base58>=2.1.0",
]
identity = [
  "PyNaCl>=1.5.0",
  "base58>=2.1.0",
]
examples = [
  "a2a-sdk[http-server]>=0.3",
]
compliance = [
  "pydantic>=2.0",
  "rfc3161ng>=2.0",
  "pyasn1>=0.4",
  "pyasn1-modules>=0.2",
  "cryptography>=41.0",
]
dev = [
  "pytest>=7.0",
  "pyyaml>=6.0",
]

[project.scripts]
a2a-exchange = "exchange.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["exchange", "compliance"]

